Fixes #23: Deleting 'pornography' and 'masturbation' behaviors
This was a user-submitted request that had coincidentally been
bothering me for a while as well. Different faster scale examples
online show slightly different behaviors. I had referenced a scale that
had the pornography and masturbation behaviors listed under the Anxiety
category. This user rightfully pointed out that these behaviors are a bit
confusing for those struggling with pornography. I've deleted them in
this commit.