Adds a FASTER category dropdown as an Email Threshold
This commit also includes fixes for an existing timezone bug (as well as
other potential bugs) related to overzealous dependency injection.

The score-based Email Threshold was removed with the score removal
changes. Replacing it was a simple boolean checkbox: send / don't send
partner emails.

This commit adds a more configurable threshold at which to send emails.