Adds category_id to user_behavior_link
Now, on a check-in, a behavior's category_id (and the category_id
associated with the selected behavior for a  processing question) is
stored alongside the other behavior data in the user_behavior_link
table. This is a helpful change that eases future development and allows
existing code to be somewhat simpler.

Also, this is a good push towards our goal of an "immutable" check-in.
Once a check-in has been done and the window for a redo is done, the
user should not be able to update any data in that check-in. The user
obviously cannot change the category a given behavior belongs to, that
relationship is hardcoded in \common\models\Category. But if that
definition were updated, behaviors that were selected in previous
check-ins would be unaltered. That's good.

Also, soon enough we will support "custom behaviors", which can be
deleted or recategorized at will. The changes in this PR will correctly
preserve history even when a custom behavior is modified.